Tests/LibWeb/Text/expected/Worker/Worker-error-event-after-gc.txt

shared worker error
worker error
